Move MFA to own directory
