+++ examples/fastspeech2/mfa/README.md
@@ -1,0 +1,1 @@
+# Extracting alignments with Montreal Forced Aligner
